Create a road and bridge network from OSM data

Creates a base graph object with the appropriate edge and vertex attributes from OSM.

konigsberg_graph(src, path_filter = automobile_highways,
  bridge_filter = all_bridges)

Arguments

src
An osmar or a base_konigsberg_graph object.

path_filter
A function that filters which Ways will be traversable in the graph. See way_filters.

bridge_filter
A function that marks which Ways are bridges that need to be crossed. See bridge_filters.

Value

A konigsberg_graph object.